CompletableFutures can now be done asynchronously on retries (#16)
Previously, if one of websocket connection, authentication, or source connection failed, then all subsequent retries had to be done synchronously with waits, but a Future obtained at any point in the stage would be updated upon success.

Now the reverse is true, i.e. only the most recent Future obtained for the request is updated to true upon success, but they retries can be performed in exactly the same asynchronous pattern as the initial attempt.