Update README.md (#52)
* Update README.md

Added an Overview section to the README, with a high-level explanation of what this project is. Added a Prerequisites section, which includes descriptions of the sample source implementation, and the full sample project zip. Added an example of a valid server.config file.

* Update README.md

Cleaned up typos

* Update README.md

One more typo.

* Update README.md

Last typo fix