Create README.md
First draft of the README.md file. 

In the Prerequisites, I wrote that there is a sample definition of the VANTIQ source implementation for the JDBC Source located under src/test/resources. Additionally, I said that there is a sample project included in the same directory. These are not currently there but I intend to add them soon. Let me know if I should remove this for now, if it is too misleading.