Add support for overriding values returned by the discovery server. (#88)
Some discovery servers appear to be misconfigured.  They return server
addresses whose host is 'localhost'.  That value is not very useful
if the connector is running on a different node.  We add two (2) ways
to fix that.

1) if replaceDiscoveredLocalhost is set to true in the source configuration,
the connector will change endpoints set to loopback addresses to point to the
same machine on which the discovery server is running.  This will make them
non-self-relative.

2) If that's not enough, set serverEndpointOverride to be the the URL to use
for the server.  After discovery, we'll replace the entire endpoint as
specified by this value.

Also:  Added additional diagnostics about discovery so that we see what
information discovery returns and how we interpret it.  Hopefully, this will
help us sort out misconfigurations more quickly.