Update connector to handle misconfigured discovered endpoints.
The OPC UA spec states that OPC Servers may be unaware of the validity of
the endpoint URLs they return.  So, clients must attempt to fix things up
upon return.

Already had code to mostly handle this, so extended it a bit. Added tests for
verification of operation.

Cleaned up list of external servers against which we validate connections.