OPC UA: Upgrade connector to a more modern Milo (SDK) version (#188)
* Update code & tests to version 0.3.* of Milo API.

Lost a little test functionality, but now fully operational.

* Update connector to handle misconfigured discovered endpoints.

The OPC UA spec states that OPC Servers may be unaware of the validity of
the endpoint URLs they return.  So, clients must attempt to fix things up
upon return.

Already had code to mostly handle this, so extended it a bit. Added tests for
verification of operation.

Cleaned up list of external servers against which we validate connections.

* Update readme with description (brief) of discovery process & assocated recovery.