Revert coord converter to be Float based as that's all it can really do.
Add capability to test for collinearity of input points.  As it turns out,
the problems extant in the GPS tests appear to have to do with the fact that
the destination points are too close together.  We succeed with we adjust
these to remove common significant digits.

Need to improve on determining whether the input is sufficient.  It's probably
doable, but need a bit more research. Might be as simple as checking it to
verify that the inputs can be converted to the outputs.  The simple
collinearity test doesn't seem sufficient.