Map coordinates in object recognition output (#190)
* Initial -- base function in rough shape -- simple tests work

* Extend testing

* Document & add note on necessary test...

* Clarify error message

* Add ORC config & tests for post processing/location mapping feature

* Add lombok plugin to gradle & cross-grade mechanism for including plugins.

* Avoid extraneous & worrisome NPEs whilst running tests

* Add _resultsAsGeoJSON_ to configuration spec & tests.

* Add copyright

* Add location mapping to connector

Enhance _ObjectRecognitionCore_ to augment results with _mappedResults_
as per Source configuration. Support results returned as straight
conversions or presented as GeoJSON.

Add tests for these.

* Add tests for GeoJson conversion.

* Clean up & clarify error re: opencv version

* Change coordinateConverter to work based on doubles vs floats.  Mostly.

* Float -> Double completion.

Still an issue that the YoloQueriesLocationMapper tests sometimes hang.  Has
to do with reconfig coming before we're paying attention to it.

* Fix race condition present with multiple tests

* Revert coord converter to be Float based as that's all it can really do.

Add capability to test for collinearity of input points.  As it turns out,
the problems extant in the GPS tests appear to have to do with the fact that
the destination points are too close together.  We succeed with we adjust
these to remove common significant digits.

Need to improve on determining whether the input is sufficient.  It's probably
doable, but need a bit more research. Might be as simple as checking it to
verify that the inputs can be converted to the outputs.  The simple
collinearity test doesn't seem sufficient.

* Add new impl of coordinate converter based on plain array math.

This allows us to use doubles rather that floats, which yields some
improved precision.  With this, our conversion of our tested GPS
conversion from online example and alleyCam1 operate as expected.

* Complete impl & test of BigDecimal-based CC

* Complete changeover to BigDecimal-based coordinate converter.

Tests, etc.

* Comment/refactoring to clean things up a bit.

* Add tests for other alley cams about which I can find data.

* Code cleanup

* Update README to describe `locationMapper` `postProcessor`.

* Add remark about 2D orientation of the location mapper.

* Reviewer comments