Fix lost message problem under heavy notification load (#200)
* Test prototype of high-load notifications

* Revert tests in wrong place

* Back to normal

* type

* Change extension source type to match the test resource given.

Really, source type *should* be created by the tests.

* Improve test diagnostics.

*MUCH* more work is needed here.

* Fix #199 -- limit the concurrency of notifications.

Notifications are async, but invisibly so. So, problematic programs
can flood the system with notifications, This will cause issues on the VANTIQ
side, but here it was causing various bits of resource exhaustion
on the websocket and/or the OkHTTP system.

We introduce a semaphore that limits the number of outstanding notifications
here. That has the effect of throttling the concurrency of messaging.

Without these changes, sending > 1000 in rapid succession would end up with
dropped messages.  With these changes, we've tested up to 150K with no
lost messages.

* Catch exception & release if send fails.