Remove need to build Open CV to build/use the Object Rec Connector (#309)
* Fix test URL for apparent new version.

* Fix typos

* Add description for configurable RTSP transports

* First pass at converting OR to use JavaCV.

This will make it easier for developers & users to build things.

* Remove dependency on OPENCV_LOC

* Test Fix

* Test & code cleanup

* Update readme

* Gradle code cleanup

* Code cleanup

* Expunge direct OpenCV calls from File & Network retriever.

Add tests for reading stream of images from file. Convert existing OpenCV calls that are problematic
on Linux to use JavaCV calls.  This helps to remove the need for OpenCV to be downloaded to the customer-
developer's machine. Camera retriever still to go.

* Restore erroneous functionality until we figure out the real issue.

* Update CameraRetriever to new structure.

Also, various bits of code cleanup & consolidation.  Centralized turning off info dump on stream open.

* Convert location mapper test to use test video file.

Fails too often due to insufficient targets in the frame.  This is one of the few tests
that's dependent upon the data in the images, so may as well convert to a known quantity.

* Avoid NPE during tests

* Remove debug stmts

* Fix excessive timeout

* Sider Fixes

* Sider Fixes

* More Sider fixes

* More Sider Fixes

* Even more sider fixes...

* Fix Sider-fix so that loopcount/frame comparison is correct

* Review comments

* Turn down detailed logging.