Use connector SDK probe capabilities (#325)
* Use declare_{healthy,unhealthy}().

We'll declare ourselves happy by default.  If the execution of some
caller code returns a MemoryError, we'll become unhealthy and let K8s
recycle us (assuming we're running in K8s).

* Update versions of connector SDK to handle cleanup more cleanly.

* Update version, fix tests

* Move to template-based generation of things containing versions.

This way, to change dependencies on other Vantiq projects' versions,
simply edit the appropriate ext-variable in the build.gradle file. The
generate*() tasks in the gradle file will regenerate the files including
this information.

* Fixup generation for cases where we're building the distribution.

Some of the tasks that have dependencies are generated later.  To get around
this, we make all the generate* tasks "go first" (make everything run after
them), and that ensures that the task dependencies are correct.

* Fixup "delete all the evidence" problem & some whitespace issues

* Allow for command-line override of version to publish

* Add document describing the publication process

* Removed -- local site instead

* Remove ref to test pypi

* Only automatically declare healthy when running in K8s.

* Update version to reflect changes in code

* Fix comment

* Update Dockerfile template for production use

* Mark forcing to debug level logging as TODO.

Can remove that after a bit of burn-in.