Add ability & tests for defining Camel endpoint's URI from the connector config.
Allows the server.config file to specify how to connect to a vantiq, and
allows vantiq endpoints to get their URLs generated appropriately.  Endpoints
with a null URI or a URI whose host is 'server.config', will have their URL
regenerated based on the data in the server.config file.  This uses the
DefaultEndpoint's (that VantiqEndpoint extends) createEndpointUri() to fill
in the details.  This is called by DefaultEndpoint() (constructor) when
the uri is null.