Add support & tests for loading routes from text.
Doing it at this level, using the camel mechanism, allows us to support any
specification types (xml, yaml, etc.) that may be added or required.  To make
use of these, the build.gradle file must include the DSL interpreters
so that we can parse & determine the components needed for the routes.  But
no code should need to be changed.