Add support for component initialization.
Some compoenents use component-specific properties to set things up. To
handle this, we support specifying the component name & the set of
properties that need initialization (and the appropriate values).  For
each some component, we'll find that component in the route, and, for
each specified property, we'll call the setter (`setPropertyName` for the
property `propertyName`) with the value provided.  We use the
class of the value provided to find the setter.  That is, if
you say you need the property `propertyName` initialized with the value
5, we will look for a method `setPropertyName(<Integer>)`.  We do not
attempt type conversions on the caller's behalf.