Make PlaceholderComponent subclass BeanComponent.
Having the `PlaceholderComponent` extend `BeanComponent` (which, in turn,
extends `DefaultComponent`) allows our placeholder to be acceptable to Camel as a bean component when it tries to "start" our discovered setup (no start really
happens).  This should permit discovery to proceed and allow things to work as
expected.

Added test based on issue report.

Fixes #372