Make PlaceholderComponent subclass BeanComponent. (#373)
* Make PlaceholderComponent subclass BeanComponent.

Having the `PlaceholderComponent` extend `BeanComponent` (which, in turn,
extends `DefaultComponent`) allows our placeholder to be acceptable to Camel as a bean component when it tries to "start" our discovered setup (no start really
happens).  This should permit discovery to proceed and allow things to work as
expected.

Added test based on issue report.

Fixes #372

* Add resolution of beans tests.

Since not starting anything, remove any hints of actual places since
that may provide a vehicle for attack. Info is public anyway, but why
copy it here.

* Dump resolved specification only under trace control.  Seems to potentially include various sensitive information