Fixes #379 (#380)
Add to list of "acceptable" errors & update list of external servers.

"Acceptable" errors are errors that indicate that the server is there but
in a broken state.  We don't want to fail our tests when some external server
is broken.

Updated list of servers against which to test. Commented out use of servers
that are not not responding (and haven't been for a while) and added a couple
more from the "publicly available servers" page.  The ones commented out
are left since they are still listed on that page, but are currently not
responding.  They are likely out of commission, but they may return. We
test against a number of external servers since they have show various
configuration or software idiosyncrasies around which we've had to code.