Add query param "consumerOutputJson" to Vantiq URL.
This instructs Vantiq consumers using this URL to exchanges on
the Camel "bus" as JSON strings.  Without this, they go on as Java Maps.

While the general way of doing this would be to marshal accordingly, allowing
this option vastly simplifies use of some kamelets without having to rewrite
the recipe.