Correctly build importable projects for assemblies from Kamelets.
Name assemblies so package name includes Camel version so they are easily
distinguishable and so that we can maintain multiple versions in our
catalog simultaneously.  As it stands, one kamelet doesn't work with version
of camel with which we build our connector. But it's time to upgrade
all that anyway & to link all the camel subprojects together w/r/t version.