Add testing in setting up source to ensure it's done.
Still seems to be a bit of flakiness whereby the source gets reconfigured for
no apparent reason.  This may be due to rules or something being added to it
or just the fact that it's been created/destroyed so many times.  But things
are better than they where -- mostly the TestJDBC#testAsync* tests that
periodically fail.