Update camel version & use the version in package names (#407)
* Update camel version to 3.21, a long term support release.

Update camelAssemblies to use same camelVersion as other connectors. Did
not start this way since kamelets didn't exist until a later version.

* Update artifact map to version 3.21.0

* Remove client close from component.  Handled by endpoint as appropriate

* Add version numbers throughout.

Also, fix some issues with config property values. Converted enum handling
to match activity pattern definitions under the plan that if we support it, it
will tend to follow the model that already exists. Will remove/enhance this
when a decision is made in the UI.