Improve handling of required properties & RAW wrapping of property values.. (#412)
Use spec/definition/required as primary indicator, but allow some properties
to have a required property that is the ultimate truth.  Ensure that this
value is set in the assembly definition so that assembly installation requires
specification of values here.

As part of assembly generation, track the properties that are marked as
passwords, credentials, or those that have UI characteristics that say they
shouldn't be displayed. For these,

* Add to their description suggesting that they be specfied as Vantiq Secrets and have their value entered as @secrets(<<secretName>>).
* Mark them in the source as needed Camel RAW wrapping.

Correspoding fixes (#409) will have the Camel connector appropriately wrapping
the property values for the Properties component.  These two actions,
in concert, allow the assembly installer to remain unaware of the RAW()
requirements.  It is still available (the user can add RAW(some value) as the
assembly config property value, but for many cases, it won't be necessary.

No issue entered.