Add ability to discover based on templates rather than (only) routes.
When importing from kamelets, dealing with templates, which is their natural
frame of mind, is easier than attempting to translate.  Found problems when the
templates included beans, as Camel tries to instantiate said Beans before
discovery (and we don't discover them anyway).

Fixed this by using the _additionalLibraries_ capability of the connector.
The locations from which the beans are loaded are typically defined in the
kamelet's list of dependencies, so we now take those into account. Before
loading the templates, we add an application class loader that includes the
additional libraries. This allows the beans to be instantiated, and everyone
remains relatively happy.

Added test for such a template.  While there, turned off the output
from ivy when we don't want it (it will remain when we have trace level logging
enabled).

This allowed us to avoid guessing where the beans might be from and including
those guesses in the connector's dependency chain (_i.e._, in gradle).