Camel Connector: handle route templates as well as routes (#414)
* Setup discovery & operation to run off of route templates as well as routes.

Fix property placeholder setting to use initialProps instead of LocalProps.
Local setting was threadLocal & didn't work with template & bean handling.

Generate routes from route templates. This requires a bit of a dance but works
better when beans are included.  Still have some work to allow dynamic
discovery of libraries required for said beans. At present, patched in
build.gradle to include known places where such beans are likely found.

* Add template test

* Upgrade camel to v3.21`

* Add ability to discover based on templates rather than (only) routes.

When importing from kamelets, dealing with templates, which is their natural
frame of mind, is easier than attempting to translate.  Found problems when the
templates included beans, as Camel tries to instantiate said Beans before
discovery (and we don't discover them anyway).

Fixed this by using the _additionalLibraries_ capability of the connector.
The locations from which the beans are loaded are typically defined in the
kamelet's list of dependencies, so we now take those into account. Before
loading the templates, we add an application class loader that includes the
additional libraries. This allows the beans to be instantiated, and everyone
remains relatively happy.

Added test for such a template.  While there, turned off the output
from ivy when we don't want it (it will remain when we have trace level logging
enabled).

This allowed us to avoid guessing where the beans might be from and including
those guesses in the connector's dependency chain (_i.e._, in gradle).

* avoid various silly actions when not at trace level logging

* Remove obsolete FIXME

* Whitespace cleanup

* Review comments