Generate route templates rather than routes (#415)
This is what's used in kamelet land, so best to keep with it.  Corresponding
change to camelConnector enables these changes.

Part of this was learning from the declared dependencies in the kamelet files.
The information there, while not necessarily complete, allows us to
pick up dependencies that might be required for beans embedded in the route
templates.

This change caused a change in behavior, at least for the aws kamelets.
Specifically, without including the declared dependencies, the AWS endpoints
were not RAW() wrapping their endpoint credentials.  With the dependency
inclusion, things are now wrapped. So, for now, I've disabled auto wrapping. It
may prove necessary for other kamelets, so the code is there, albeit commented
out. If we need to  add it back conditionally, we can.

This situation is curious in that the camel aws* components _should_ include
whatever dependencies they need in the their respective pom.xml files, so that
dependency management can take care of such things. But it appears not to,
so handling this changed things.