Fix CamelComponent issue with structured messages via Google PubSub (at least) (#417)
* Fixes #416

When dealing with (at least) message from google pubsub, the header values are
not understood by the Jackson object mapper.  Instead, we'll revert to simply
traversing the case insensitive structure given, converting it into a regular
hashmap before sending it back to vantiq.  This is only an issue when asking
for structured messages.

* Minor update to handle other types...

* Handle most cases simply