Handle deserialization of byte arrays when given those from someplace.
GooglePubSub seems to like to hand out byte arrays, leaving it to the recipient
to deserialize them.  Handle conversion to/from JSON (which is generally what
Vantiq can use and/or will send), but if that fails, fall back to low-level
Java network deserialization.

Once that's done, if possible, convert to map for Vantiq system consumption.