Streamline processing of input streams for the producer.
Since we do that when it happens, read the bytes from the stream then
deal with deserializing what's there.  Otherwise, a bunch of repeated
code.

Enhance tests to deal with _consumerOutputJson_'s now streaming output.  That's
what marshaling does, so we should match that behavior to fit into the
undocumented ecosystem.