Stream output from consumerOutputJson -- for Kinesis & other support (#429)
* Produce & consume byte streams when appropriate.

This is still iffy for the consumer -- we now produce an InputStreamCache(d)
when doing JSON output.  This works with Kinesis, but not clear if that's true
everywhere. However, it approximately matches what marshal/json/jackson does,
so that or something like it is probably the right thing.

* Cleanup log4j2.properties files

* Streamline processing of input streams for the producer.

Since we do that when it happens, read the bytes from the stream then
deal with deserializing what's there.  Otherwise, a bunch of repeated
code.

Enhance tests to deal with _consumerOutputJson_'s now streaming output.  That's
what marshaling does, so we should match that behavior to fit into the
undocumented ecosystem.

* Deprecate `consumerOutputJson` in favor of `consumerOutputJsonStream`.

Converted to use ...Stream to more closed emulate a marshal step, improving
interoperability with other Camel Components (since that's what many other
components expect).  Deprecated extant `consumerOutputStream` in case any
existing uses exist.

Added tests for both endpoint parameters, verifying correct operaiton of both.