Fix schema impedance mismatch kamelets vs. assemblies
Kamelets allow parameters to be required AND have a default value.  Vantiq
assembly configuration properties (generated from the Kamelet parameters) do
not.  So, if we have something that's required but we find a default value,
suppress the _required_ property -- the provision of the default value in the
assembly will suffice