Partial attempt to reduce OR failures in extension sources (#452)
* Move camera check util to more available location

* Check on camera availability at test start.

This has a whole wherein if the camera is being sporadic, some tests may still
fail (or all tests will get skipped).  However, checking at each test 1) doesn't
really improve things that much, and 2) makes the tests take forever if the
camera address is timing out.

So we compromise

* Cycle Camera in use to find one that works.

Not sure how well this will work when dark -- may have to have a strategy
for determining whether there tend to be identifiable objects in images.

Leaving that for a later date.