Fixes #451
Improves description & visibility thereof for information about assemblies.

High level description now includes the source name (since that's not visible)
and the descriptive information from the kamelet.  That information is also
in the overview document for reference post install.

The overview documents have been added to the visible resources (previously,
that didn't seem to be an issue but something's changed since I last checked).

(These were suggestions from Mark as the initial reviewer.)