Fix issue with Camel routes using choice stmts and/or properties references (#457)
* Use supported mechanism for setting properties as required.

* Fix discovery so that it can recover from discovered routes containing Choice

CamelContext used for discovery was being closed.  However, once the route is
read, some expressions in the Choice stmts remain with pointers to old context.
Seems like this should be cleaned up, but it's not.  So we'll leave
the old context open.

* Add test using FHIR server to verify routes with Choice stmts operate.

* Bump image tag for connector since we've made some changes.

Move default up to new version so that all assemblies (we've tested) will us
the new version when deploying things.

* self review

* Review comments

* Remove unnecessary setting in test