Add support for route overrides -- and use one for the fair-sink (#458)
* Allow ability to provide camel route documents.

Useful when we want to adjust what the kamelet-esque thing does or when there
are issues with the route document as generated.

* Add replacement route for the fhir-sink assembly.

As currently generated (as extracted from the kamelet in question), there
is an endpoint setting that doesn't seem to exist.  With this modified
route document, we remove that error.

* Add notes document describing usage of FHIR&reg;-sink assembly.

* Override route.

This adds a roundtrip so that results can be fed back to the Vail application
if desired.  PUBLISH will ignore them, SELECT will see them.

Probably "knows" more than it should in that the Vail user must query against
the underlying source.

* Obviate requirement to do `import` & `publishAssemblies` separately.

Can now handle with `assemble`, etc.

* review comments

* Fix issue for builds where import&publish conditions checked at config time.

Should be at run time -- move stuff into `doLast{}` closure