Add verified support for Camel JIRA assemblies (#463)
* Add ability to override source configuration.

Useful when we need to add other libraries or something.

* Rev 1 of jira-source override.

Not quite right, I don't think.  But there appear to be somewhat well-known
issues with com.atlassian.sal:sal-api:4.4.2 not resolving in various forms.
Still trying to track down what's awry there (other than it appears that
Atlassian may have yanked support).  Added test sans Kamelet (just
using the jira component with Camel 3.21 and nothing resolves.

More investigation to do.

* Add override & test for errors during camel-jira resolution.

There seems to be an issue with the camel-jira component where it
requires (as per pom.xml) a maven plugin that's not available (and not
really required).  We'll now ignore that.

We further ensure that a route containing jira loads properly.

* Update atlassian repo for jira source

* Clean up debugging msgs

* Update docker image version for camel connector

* Handle resolution of jira stuff to handle errors

* Add JIRA test to verify sal-api substitution

* Add ability to check on errors starting up.

Mostly useful in tests.

* Add mediator & type overrides so that we can compensate for maven issues.

Some of the things used by the atlassian/jira stuff seems to be registered
incorrectly, so we need to be able to work around those things.  We supply
the generic ability to provide overrides for the type for particular
library dependendies, then override the artifact(s) required as appropriate.

* Add test for JIRA startup to verify our mediation capabilities.

Not really testing JIRA component;  not really our job, though we may
offer some capabilities for such verification based on gradle property setup.

* Upgrade jira test so that it can be configured via system properties.

Make associated changes to build.gradle so that it will set these system
properties for test tasks.

* Add source overrides for other JIRA-esque assemblies

* Update image tag for camel connector

* Comment cleanup

* Fix comment going nowhere.