Upgrade Camel -> version 4 for CamelComponent & CamelConnector (#512)
* Upgrade builds to camel 4 & complete update to new lombok plugin

* Camel Component code -> Camel 4.x, tests OK

* Update camelConnector to JDK 17 compatibility

* code cleanup

* Standardize jackson version to use

* more standardization

* Updates for camel 4, jdk 17, junit 5

* Update artifacts to camel 4.4.3

* Update versions & plugin

* Upgrade internal support.

Also, due to some idiocsyncracies in the Jira repo, change behavior in
error condition where things don't resolve. We'll log warnngs, but this
allows customer to add _additionalLibraries_ to pick up dependencies that
are required but (apparently) misconfigured elsewhere.

* Add (albeit commented out) tracing for Ivy.

Useful sometimes in debugging dependency resolution, so leaving it here
in case others need it.

* Upgrade to junit5

Also, deal with ramifications of Jdk 17 with Jackson for dealing with
Optionals.  Purportedly, this is fixed but I suspect Camel's not setting it
up correctly.  Since we don't really care, we just avoid the issue since
it's not adding anything to our tests.

* Cleanup

* More pre-review cleanup

* Remove commented out logging.

Was left in pending completion of debug, but no need to keep it further.