Update assembly generation for Camel v4 (#516)
* Ignore parameters of the form `inBody: ...`

These are really more human information (_put the resource in the body of
the message_).  The camel route reader chokes on them.

* Upgrade override files for Camel 4.

fhir -- names of parameters/properties changed.
jira -- need to add some _additionalLibraries_ as there are some apparent
errors in the atlassian repo.  Libraries missing versions, etc.

* Add camel4connector assembly -- for use with camel 4.

Adding this allows use of the v3... assemblies to coexists with v4...

* Build & use the camel4connector assembly

* Correctly handle merging of additional libraries

* Add log4J/slf4j implementations to the base build (rather than test).

This includes the logger in the camelConnector so the logs are correctly
generated & visible.

* Fix up template reference post name change

* Fixup references so that interface services are correctly recognized.

* Base docker image of Java 17 corretto

* Update readme with package names for Camel version 4 assemblies

* Fix outdated reference