Upgrade Camel component, connector, and assemblies to Camel 4 (#517)
* Comment for attribution...

No substantive change

* Upgrade Camel -> version 4 for CamelComponent & CamelConnector (#512)

* Upgrade builds to camel 4 & complete update to new lombok plugin

* Camel Component code -> Camel 4.x, tests OK

* Update camelConnector to JDK 17 compatibility

* code cleanup

* Standardize jackson version to use

* more standardization

* Updates for camel 4, jdk 17, junit 5

* Update artifacts to camel 4.4.3

* Update versions & plugin

* Upgrade internal support.

Also, due to some idiocsyncracies in the Jira repo, change behavior in
error condition where things don't resolve. We'll log warnngs, but this
allows customer to add _additionalLibraries_ to pick up dependencies that
are required but (apparently) misconfigured elsewhere.

* Add (albeit commented out) tracing for Ivy.

Useful sometimes in debugging dependency resolution, so leaving it here
in case others need it.

* Upgrade to junit5

Also, deal with ramifications of Jdk 17 with Jackson for dealing with
Optionals.  Purportedly, this is fixed but I suspect Camel's not setting it
up correctly.  Since we don't really care, we just avoid the issue since
it's not adding anything to our tests.

* Cleanup

* More pre-review cleanup

* Remove commented out logging.

Was left in pending completion of debug, but no need to keep it further.

* Update assembly generation for Camel v4 (#516)

* Ignore parameters of the form `inBody: ...`

These are really more human information (_put the resource in the body of
the message_).  The camel route reader chokes on them.

* Upgrade override files for Camel 4.

fhir -- names of parameters/properties changed.
jira -- need to add some _additionalLibraries_ as there are some apparent
errors in the atlassian repo.  Libraries missing versions, etc.

* Add camel4connector assembly -- for use with camel 4.

Adding this allows use of the v3... assemblies to coexists with v4...

* Build & use the camel4connector assembly

* Correctly handle merging of additional libraries

* Add log4J/slf4j implementations to the base build (rather than test).

This includes the logger in the camelConnector so the logs are correctly
generated & visible.

* Fix up template reference post name change

* Fixup references so that interface services are correctly recognized.

* Base docker image of Java 17 corretto

* Update readme with package names for Camel version 4 assemblies

* Fix outdated reference