Improve ObjectRec test reliability & consistency (#524)
* Extend camera connectivity to optionally allow test to fail.

* Add test to fail if no IP cameras found.

Use trial RTSP camera to test rtsp net retriever.  Old RTSP test is not working.

* Move valid camera test to be inside junit infra so tests will be skipped.

* Clean up camera list.

Local traffic camera opens (that is, can connect to it), but then fails to
to return things.

* Enhance network retriever tests to get get a few images.

We are changing the other tests to be file-based (for reliability of the test),
so we should test this a tiny bit more thoroughly.  Change the test to
fetch multiple images & verify (md5 digest) and verify that they aren't all
the same image.

* Update the YOLO & other OR connector tests to use file-based video.

Online camera reliability is getting suspect (these are all random,
unmaintained things), and using a network camera isn't really enhancing the
things being tested.  Go for something more reliable.

* Update threshold used fpr the precropping tests.

What we find there seems to be adversely affected by machine speed, and
these tests were failing (lately) more often than not.  One of the objects
found varied (in confidence) between 57 & 61.  With the threshold set to
60, it disappeared sporacically.

Fixing this improves the reliability w/o changing the test efficacy.

* One more fix.