Add timeout for socket connection.
During tests, this is used to probe the various test servers we use to
attempt to verify that that they are around at the time of the test. Most
of them are servers offered out of courtesy to facilitate interoperability
testing, which is what we're doing.  Without the timeout setting, each
timeout is taking 2 minutes to complete and tests run for quite a while.

With the change, tests down to a few minutes altogether.