Few more fixups for buildDir
The property buildDir is deprecated.  To get the value, the "correct"
mechanism is "layout.buildDirectory".  But that returns some object, not
a string.  In most cases it gets converted and works find.  But in the OR
project, it fails to convert so you get serialization-style errors on
SystemProperties.  No idea why that's different.

This changes the calls to be cast to strings.  Which operates as expected.