Uptake Vantiq dcompose & upgrade to gradle 8 (#527)
* fixup deprecated gradle properties

* Upgrade to gradle 8.9

* Fix a bunch of deprecations for gradle 8.x & prep for gradle 9

* Fix a few more complaints

* Add timeout for socket connection.

During tests, this is used to probe the various test servers we use to
attempt to verify that that they are around at the time of the test. Most
of them are servers offered out of courtesy to facilitate interoperability
testing, which is what we're doing.  Without the timeout setting, each
timeout is taking 2 minutes to complete and tests run for quite a while.

With the change, tests down to a few minutes altogether.

* Few more fixups for buildDir

The property buildDir is deprecated.  To get the value, the "correct"
mechanism is "layout.buildDirectory".  But that returns some object, not
a string.  In most cases it gets converted and works find.  But in the OR
project, it fails to convert so you get serialization-style errors on
SystemProperties.  No idea why that's different.

This changes the calls to be cast to strings.  Which operates as expected.

* Fix G-String spec

* More cleanup of build directory stuff

New mechanism is very cumbersome compared to old.  And seemingly a bit tweeky
on different platforms.

* review comments