Add test for batch delete.
This adds that test and, by virtue of that, deletes all the junk
that the other bundle tests add in.  Required some amount of work track
the placeholder URL -> actual URL mapping, since the documents
contain the placeholders but delete's have to be done on the actual values.

In real life, we just need the actual so we can delete those.

This impacted downstream tests as they didn't have to track the "bat girl"
additions any more.