Merge FHIR Staging into master (#568)
* Set up FHIR server for testing (#542)

* Create base test environment & server.

* Script to load sample data into FHIR serverr

* Move loading script into gradle file.

Add build.gradle & add module into overall project.
Add download & load-to-server of sample data.`

* Remove shell script.

Problematic on windows, so do it all in build.gradle file

* Add Accept header to request.

If you don't specify the Accept header (as app/json or, presumably, app/xml),
it returns HTML & doesn't actually insert the data.  Since we're trying to load
data, this is a problem.

* Make database in build area so it goes away.

Also, properly name postgres container.

* Add postgres versions of fhir server

mostly in case we need them.

* Add dockerized fhir server using in-memory database.

That should be sufficient for testing.

* Simplify docker compose.

Matches what'll be in the build.gradle file.

* Add startup, shutdown, and provisioning of HAPI FHIR server.

We start/load/<run tests>/stop the server via docker. Our server has no
long-term storage, so we can monkey with the contents during the tests
and not worry about effects later.

* Remove comment that complained about pilot error.

* Remove old junk included via copy & paste.

* remove doubled line

* Add FHIR instance operations (#549)

* Add some basic tests.

Obviously more to do...

* Add some batman test data...

* Add Vantiq project that is the FHIR assembly

* Add searchType() procudure to service.

This will be a precursor to the actual search that will be added in a subsequent
PR.  However, we need it now because resource ids assigned during import
are not consistent (at least across machines, probably not in general), so
to do a "read" operation (which requires an id) we have to know the id.

* Fixup comment

* Expand tabs

* Review comments

* Add search capabilities for types. (#550)

* Add search capability for types.

Update test accordingly.

* Add (really) search capability for types.

Update test accordingly.

* Fix typo

* Update assembly adding search & link/paging capabiities

* Add search tests for unrestricted cases

* Add testing for qualifications on searches

* Update assembly to handle generalParameters/modifiers.

Add & commonize handling where appropriate, remove where inappropriate.

Some general cleanup

* Add tests for general parameters/modifiers

* Clean up initialization code

* Fix typo

* Fix copied comment

* Review comments.

Mostly moving comments to absolute top of file.  A few other minor changes

* Add support for conditional create/update/delete (#553)

* Redo of interaction operations to more fully encompass the standard.

Expand general parameters to include headers.  Return a FHIR Response type
that includes headers & status.

* Add types in support of passing headers & returning status

* Update tests to handle new structures passed in & returned

* Properly push conditionals in headers thru update & create

* Add tests for conditional create & update.

Update is now ignored since the HAPI server doesn't seem to support it.

* Add assembly config parameter for Http method for search.

Add tests for said configurations & methods.

* EXTPSDK: resolve CVE/Dependabot issue re: jinja2 (#556)

* Update jinja2 version

* Update pypi version number

* PythonExecSource: Update jinja2 dependency/ (#557)

* Update jinja2 version as per dependabot.

Update other, related, Vantiq dependencies.

* Update package version

* Fix formatting

* Remove leftover debug stuff

* Add final (I think) set of FHIR interactions to our FHIR adaptation. (#559)

* Add patch, history, & bundle interaction capabilities

* Add test data for bundle interactions

* Add tests for patch, bundle/batch, and bundle/transaction interactions

* Remove entries that's harder to track.

It's not testing anything here.

* Add test for batch delete.

This adds that test and, by virtue of that, deletes all the junk
that the other bundle tests add in.  Required some amount of work track
the placeholder URL -> actual URL mapping, since the documents
contain the placeholders but delete's have to be done on the actual values.

In real life, we just need the actual so we can delete those.

This impacted downstream tests as they didn't have to track the "bat girl"
additions any more.

* Add tests for the variants on history interactions

* Self review/comment cleanup & a couple of renamed variables.

* Add support for Basic & OAuth authentication (#564)

* Add setup & handling for none, basic, and oauth authentication

* Update tests to include changes for auth'n

* Change oauth configs to work within Vantiq requirements

* Add & improve tests for auth'n

* Expunge undocumented stuff from procedure declarations.

Clean up comments a bit.

* Add documentation to FHIR Assembly (#566)

* FHIR Assembly doc -- initial

* Add note about TBD for building...

* Add description to service

* Add development/publishing instructions

* Update build.gradle with publishing capabilities.

Adapted from CamelAssemblies

* Review comments/properly escape angle brackets in text.

* Replace old & incomplete phraseology